+++ stars-data-av/src/main/kotlin/tools/aqua/stars/data/av/dataclasses/VehicleType.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023-2024 The STARS Project Authors
+ * Copyright 2024 The STARS Project Authors
  * SPDX-License-Identifier: Apache-2.0
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -17,23 +17,12 @@
 
 package tools.aqua.stars.data.av.dataclasses
 
-import kotlin.math.pow
-import kotlin.math.sqrt
-
-/**
- * Data class for 3D locations.
- *
- * @property x The x ordinate.
- * @property y The y ordinate.
- * @property z The z ordinate.
- */
-data class Location(val x: Double, val y: Double, val z: Double) {
-  companion object {
-    /**
-     * Calculates the Euclidean distance between two locations, i.e., the square root of the sum of
-     * the squared ordinates.
-     */
-    fun euclideanDistance(loc1: Location, loc2: Location): Double =
-        sqrt((loc1.x - loc2.x).pow(2) + (loc1.y - loc2.y).pow(2) + (loc1.z - loc2.z).pow(2))
-  }
+/** Enum class for vehicle types present in CARLA. */
+enum class VehicleType {
+  CAR,
+  TRUCK,
+  VAN,
+  BUS,
+  MOTORCYCLE,
+  BICYCLE
 }